+++ CHANGELOG.md
@@ -1,2 +1,29 @@
+## 0.0.10
+- Updated eiffel-remrem-parent version
+
+## 0.0.9
+- Migrated from gradle to maven
+
+## 0.0.8
+- Updated GSON version
+
+## 0.0.7
+- Added method to get template for event types.
+
+## 0.0.6
+- Added method to get supported event types.
+
+## 0.0.5
+- Removed methods to get event family, type and domainId.
+
+## 0.0.4
+- Added method to get event type.
+
+## 0.0.3
+- Added method to get Routing key.
+
+## 0.0.2
+- Added copyright headers to the source code.
+
 ## 0.0.1
-- Added  
+- Added